rename: 토큰 처리 인터페이스와 토큰 타입 enum을 implement 계층으로 이동
