+++ src/main/kotlin/com/yourssu/scouter/common/implement/domain/authentication/TokenType.kt
@@ -1,0 +1,7 @@
+package com.yourssu.scouter.common.implement.domain.authentication
+
+enum class TokenType {
+    ACCESS,
+    REFRESH,
+}
+
