--- src/main/kotlin/com/yourssu/scouter/common/implement/support/security/token/TokenType.kt
@@ -1,7 +1,0 @@
-package com.yourssu.scouter.common.implement.support.security.token
-
-enum class TokenType {
-    ACCESS,
-    REFRESH,
-}
-
